Review cleanup and improvements.

+++ hack/v1migration/phase2.sh
@@ -10,8 +10,6 @@ fi
 WORKDIR=$1
 echo "Using workdir: $WORKDIR"
 
-oc project hive
-
 oc scale -n hive deployment.v1.apps/hive-operator --replicas=0
 oc scale -n hive deployment.v1.apps/hive-controllers --replicas=0
 oc scale -n hive deployment.v1.apps/hiveadmission --replicas=0
@@ -22,27 +20,24 @@ then
     exit 1
 fi
 
+echo "WARNING: this script will delete all Hive custom resources and their definitions"
+echo "It should only be used during migrations to the v1 API."
+read -r -p "Do you wish to proceed? [y/N] " response
+if [[ ! "$response" =~ ^([yY][eE][sS]|[yY])$ ]]
+then
+	exit 1
+fi
 
-HIVE_TYPES=( checkpoints clusterdeployments clusterdeprovisionrequests clusterimagesets clusterprovisions clusterstates dnsendpoints dnszones hiveconfig selectorsyncidentityprovider selectorsyncset syncidentityprovider syncsetinstance syncset )
+readarray -t HIVE_TYPES <<< \
+"$(oc get crd -o json | \
+jq -r '.items[] | select(.spec.group="hive.openshift.io") | select(.spec.version="v1alpha1") | .spec.names.plural')"
+
 for i in "${HIVE_TYPES[@]}"
 do
 	:
 	echo "Deleting all ${i}"
 	oc delete ${i}.hive.openshift.io -A --all
+	echo "Deleting ${i}.hive.openshift.io CRD"
+	oc delete customresourcedefinition.apiextensions.k8s.io/${i}.hive.openshift.io
 done
-
-oc delete customresourcedefinition.apiextensions.k8s.io/checkpoints.hive.openshift.io
-oc delete customresourcedefinition.apiextensions.k8s.io/clusterdeployments.hive.openshift.io
-oc delete customresourcedefinition.apiextensions.k8s.io/clusterdeprovisionrequests.hive.openshift.io
-oc delete customresourcedefinition.apiextensions.k8s.io/clusterimagesets.hive.openshift.io
-oc delete customresourcedefinition.apiextensions.k8s.io/clusterprovisions.hive.openshift.io
-oc delete customresourcedefinition.apiextensions.k8s.io/clusterstates.hive.openshift.io
-oc delete customresourcedefinition.apiextensions.k8s.io/dnsendpoints.hive.openshift.io
-oc delete customresourcedefinition.apiextensions.k8s.io/dnszones.hive.openshift.io
-oc delete customresourcedefinition.apiextensions.k8s.io/hiveconfigs.hive.openshift.io
-oc delete customresourcedefinition.apiextensions.k8s.io/selectorsyncidentityproviders.hive.openshift.io
-oc delete customresourcedefinition.apiextensions.k8s.io/selectorsyncsets.hive.openshift.io
-oc delete customresourcedefinition.apiextensions.k8s.io/syncidentityproviders.hive.openshift.io
-oc delete customresourcedefinition.apiextensions.k8s.io/syncsetinstances.hive.openshift.io
-oc delete customresourcedefinition.apiextensions.k8s.io/syncsets.hive.openshift.io
